requirements.txt
streamlit
requests
pillow
pyzbar
cassandra-driver
openai